Merge in SSBD2023/ssbd202302 from bugfix/SSBD202302-530-edytowac-przycisk-usuniecia-oceny to master

* commit '3dea48907a482bc2e9fbb67642b0ccde3f084ba1':
  SSBD202302-530 Correct front

+++ src/main/ssbd202302-frontend/src/app/pages/client-rates-page/client-rates-page.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnDestroy, OnInit} from '@angular/core';
 import {animate, state, style, transition, trigger} from "@angular/animations";
-import {Subject, takeUntil, tap} from "rxjs";
+import {first, Subject, takeUntil, tap} from "rxjs";
 import {MatTableDataSource} from "@angular/material/table";
 import {ProductService} from "../../services/product.service";
 import {NavigationEnd, Router} from "@angular/router";
@@ -11,6 +11,7 @@ import {HttpErrorResponse} from "@angular/common/http";
 import {TranslateService} from "@ngx-translate/core";
 import {AlertService} from "@full-fledged/alerts";
 import { EventEmitter } from '@angular/core';
+import {DialogService} from "../../services/dialog.service";
 
 @Component({
   selector: 'app-client-rates-page',
@@ -54,7 +55,8 @@ export class ClientRatesPageComponent implements OnInit, OnDestroy {
     private rateService: RateService,
     private router: Router,
     private translate: TranslateService,
-    private alertService: AlertService
+    private alertService: AlertService,
+    private dialogService: DialogService
   ) {}
 
   ngOnInit(): void {
@@ -202,24 +204,37 @@ export class ClientRatesPageComponent implements OnInit, OnDestroy {
   }
 
   removeRating(orderProduct: OrderProductWithRate): void {
-    orderProduct.rate = 0;
-    this.rateService.removeRate(orderProduct.product.productGroup.id)
+    this.translate
+      .get('dialog.remove.rate')
       .pipe(takeUntil(this.destroy))
-      .subscribe({
-        next: (rate: Rate) => {
-          this.loading = false;
-          this.loadClientProducts();
-        },
-        error: (e: HttpErrorResponse) => {
-          this.loading = false;
-          orderProduct.rate = orderProduct.oldRate;
-          this.translate
-            .get(e.error.message || 'exception.moz.rate.notfound')
-            .pipe(takeUntil(this.destroy))
-            .subscribe((msg) => {
-              this.alertService.danger(msg);
-            });
-        },
+      .subscribe((msg) => {
+        const ref = this.dialogService.openConfirmationDialog(msg, 'primary');
+        ref
+          .afterClosed()
+          .pipe(first(), takeUntil(this.destroy))
+          .subscribe((result) => {
+            if (result === 'action') {
+              orderProduct.rate = 0;
+              this.rateService.removeRate(orderProduct.product.productGroup.id)
+                .pipe(takeUntil(this.destroy))
+                .subscribe({
+                  next: (rate: Rate) => {
+                    this.loading = false;
+                    this.loadClientProducts();
+                  },
+                  error: (e: HttpErrorResponse) => {
+                    this.loading = false;
+                    orderProduct.rate = orderProduct.oldRate;
+                    this.translate
+                      .get(e.error.message || 'exception.moz.rate.notfound')
+                      .pipe(takeUntil(this.destroy))
+                      .subscribe((msg) => {
+                        this.alertService.danger(msg);
+                      });
+                  },
+                });
+            }
+          });
       });
   }
